Removed IConfigManager, split /ftb edit_config to use <ID>, moved config editing from server side to API, WIP config manager rewrite

+++ src/main/java/com/feed_the_beast/ftbl/gui/GuiIntField.java
@@ -76,7 +76,7 @@ public class GuiIntField extends GuiLM
             }
         };
 
-        textBox.setText(Double.toString(def));
+        textBox.setText(Integer.toString(def));
         textBox.textRenderX = -1;
         textBox.textRenderY = 6;
         textBox.textColor = 0xFFEEEEEE;
